refactor(app-core): split the node-bound MemryApp cluster into apps/cli, keep the package platform-free
Spec 001-mobile-app T021. apps/cli was the only consumer of createMemryApp —
desktop imports just /reminders, /markdown, /ids and mobile just /ids — so the
node-touching implementations move to their consumer instead of hiding behind
seams mobile will never call: agent, database, paths, folder-view, folders,
inbox, locale, note-files, notes, properties, sync, templates, versions and
memry-app land in apps/cli/src/app-core/ as import-path-only diffs, along with
paths/memry-app tests and the better-sqlite3-backed bookmarks/reminders suites
(they need a node runtime; the services they test stay in the package).

What stays in @memry/app-core is platform-free: ids, markdown, reminders,
bookmarks, calendar, graph, saved-filters, search-tools, settings, tags, tasks,
plus a new service-types.ts holding the NotesService/InboxService/
TemplatesService contracts (extracted so pure modules and later mobile code can
depend on the contracts without reaching node implementations; the moved impls
re-export them, so their public surface is unchanged). The pure services'
DataDb narrows no further: it now aliases the driver-agnostic DrizzleDb from
@memry/db-schema/drizzle-db, same widening as the sync extraction. The root
export of @memry/app-core becomes the contract surface; a ./* subpath export
serves the rest (mobile-facing ids/markdown/reminders exports unchanged).

apps/cli takes over the node deps (better-sqlite3, drizzle better-sqlite3
driver, gray-matter, marked, storage/domain packages); app-core drops them.
Also folds runtime.ts's RuntimeSyncDb intersection into its single remaining
constituent — the two halves became the same widened type in the extraction
and fresh-cache eslint (no --cache masking) flags the duplication as an error.

Verification: root typecheck 18/18 green, app-core + cli node:test suites pass
(15/15 in cli incl. real-DB bookmarks/reminders), architecture boundary check
green, eslint 0 errors.
