+++ apps/cli/src/app-core/inbox.ts
@@ -1,137 +1,13 @@
+import type { InboxRecord, InboxBulkResponse, InboxService } from '@memry/app-core/service-types'
+export type { InboxRecord, CaptureTextInput, CaptureLinkInput, CaptureFileInput, UpdateInboxInput, InboxFileResponse, InboxBulkResponse, InboxFilingHistoryEntry, InboxCapturePattern, InboxService } from '@memry/app-core/service-types'
 import fs from 'node:fs/promises'
 import path from 'node:path'
 import { and, asc, eq } from 'drizzle-orm'
 import { inboxItems, inboxItemTags, settings as settingsTable } from '@memry/db-schema/data-schema'
 import type { DataDb } from './database.ts'
-import { createId } from './ids.ts'
+import { createId } from '@memry/app-core/ids'
 import type { NotesService } from './notes.ts'
-import type { TasksService } from './tasks.ts'
-
-export interface InboxRecord {
-  id: string
-  type: string
-  title: string
-  content: string | null
-  createdAt: string
-  modifiedAt: string
-  filedAt: string | null
-  filedTo: string | null
-  filedAction: string | null
-  archivedAt: string | null
-  viewedAt: string | null
-  snoozedUntil: string | null
-  snoozeReason: string | null
-  processingStatus: string | null
-  sourceUrl: string | null
-  sourceTitle: string | null
-  metadata: unknown | null
-  attachmentPath: string | null
-  thumbnailPath: string | null
-  tags: string[]
-}
-
-export interface CaptureTextInput {
-  title?: string
-  content: string
-  tags?: string[]
-}
-
-export interface CaptureLinkInput {
-  url: string
-  tags?: string[]
-}
-
-export interface CaptureFileInput {
-  filePath: string
-  mimeType?: string
-  title?: string
-  tags?: string[]
-}
-
-export interface UpdateInboxInput {
-  title?: string
-  content?: string | null
-}
-
-export interface InboxFileResponse {
-  success: boolean
-  filedTo: string | null
-  noteId?: string
-  taskId?: string
-  error?: string
-}
-
-export interface InboxBulkResponse {
-  success: boolean
-  processedCount: number
-  errors: Array<{ itemId: string; error: string }>
-}
-
-export interface InboxFilingHistoryEntry {
-  id: string
-  itemId: string
-  itemType: string
-  itemTitle: string
-  filedTo: string
-  filedAction: string
-  filedAt: string
-  tags: string[]
-}
-
-export interface InboxCapturePattern {
-  timeHeatmap: number[][]
-  typeDistribution: Array<{ type: string; count: number; percentage: number; trend: 'stable' }>
-  topDomains: Array<{ domain: string; count: number }>
-  topTags: Array<{ tag: string; count: number }>
-}
-
-export interface InboxService {
-  captureText(input: CaptureTextInput): Promise<InboxRecord>
-  captureLink(input: CaptureLinkInput): Promise<InboxRecord>
-  captureFile(input: CaptureFileInput): Promise<InboxRecord>
-  get(id: string): Promise<InboxRecord | null>
-  list(options?: {
-    includeArchived?: boolean
-    includeSnoozed?: boolean
-  }): Promise<{ items: InboxRecord[]; total: number }>
-  tags(): Promise<Array<{ tag: string; count: number }>>
-  stats(): Promise<{
-    totalItems: number
-    archivedCount: number
-    snoozedCount: number
-    viewedCount: number
-    itemsByType: Record<string, number>
-  }>
-  archived(options?: {
-    search?: string
-    limit?: number
-    offset?: number
-  }): Promise<{ items: InboxRecord[]; total: number; hasMore: boolean }>
-  filingHistory(options?: { limit?: number }): Promise<{ entries: InboxFilingHistoryEntry[] }>
-  patterns(): Promise<InboxCapturePattern>
-  getStaleThreshold(): Promise<number>
-  setStaleThreshold(days: number): Promise<{ success: boolean }>
-  update(id: string, input: UpdateInboxInput): Promise<InboxRecord>
-  archive(id: string): Promise<InboxRecord>
-  unarchive(id: string): Promise<InboxRecord>
-  convertToNote(id: string): Promise<InboxFileResponse>
-  convertToTask(id: string): Promise<InboxFileResponse>
-  linkToNote(
-    id: string,
-    noteId: string,
-    tags?: string[]
-  ): Promise<{ success: boolean; error?: string }>
-  snooze(id: string, until: string, reason?: string): Promise<InboxRecord>
-  unsnooze(id: string): Promise<InboxRecord>
-  snoozed(): Promise<{ items: InboxRecord[]; total: number }>
-  bulkArchive(ids: string[]): Promise<InboxBulkResponse>
-  bulkSnooze(ids: string[], until: string, reason?: string): Promise<InboxBulkResponse>
-  bulkTag(ids: string[], tags: string[]): Promise<InboxBulkResponse>
-  deletePermanent(id: string): Promise<boolean>
-  addTag(id: string, tag: string): Promise<InboxRecord>
-  removeTag(id: string, tag: string): Promise<InboxRecord>
-  markViewed(id: string): Promise<InboxRecord>
-}
+import type { TasksService } from '@memry/app-core/tasks'
 
 function nowIso(): string {
   return new Date().toISOString()
