+++ apps/cli/src/app-core/memry-app.test.ts
