+++ apps/cli/src/app-core/memry-app.ts
@@ -1,17 +1,17 @@
 import type { OpenedDatabases } from './database.ts'
 import type { AgentService } from './agent.ts'
 import { createAgentService } from './agent.ts'
-import type { BookmarksService } from './bookmarks.ts'
-import { createBookmarksService } from './bookmarks.ts'
-import type { CalendarService } from './calendar.ts'
-import { createCalendarService } from './calendar.ts'
+import type { BookmarksService } from '@memry/app-core/bookmarks'
+import { createBookmarksService } from '@memry/app-core/bookmarks'
+import type { CalendarService } from '@memry/app-core/calendar'
+import { createCalendarService } from '@memry/app-core/calendar'
 import { openDatabases } from './database.ts'
 import type { FoldersService } from './folders.ts'
 import { createFoldersService } from './folders.ts'
 import type { FolderViewService } from './folder-view.ts'
 import { createFolderViewService } from './folder-view.ts'
-import type { GraphService } from './graph.ts'
-import { createGraphService } from './graph.ts'
+import type { GraphService } from '@memry/app-core/graph'
+import { createGraphService } from '@memry/app-core/graph'
 import type { InboxService } from './inbox.ts'
 import { createInboxService } from './inbox.ts'
 import type { LocaleService } from './locale.ts'
@@ -34,24 +34,24 @@ import { createNotesService } from './notes.ts'
 import type { PropertiesService } from './properties.ts'
 import { createPropertiesService } from './properties.ts'
 import { ensureVaultLayout, writeVaultConfig, type VaultConfig } from './paths.ts'
-import type { RemindersService } from './reminders.ts'
-import { createRemindersService } from './reminders.ts'
-import type { SettingsService } from './settings.ts'
-import { createSettingsService } from './settings.ts'
-import type { SavedFiltersService } from './saved-filters.ts'
-import { createSavedFiltersService } from './saved-filters.ts'
-import type { SearchReasonsService, SearchStats } from './search-tools.ts'
+import type { RemindersService } from '@memry/app-core/reminders'
+import { createRemindersService } from '@memry/app-core/reminders'
+import type { SettingsService } from '@memry/app-core/settings'
+import { createSettingsService } from '@memry/app-core/settings'
+import type { SavedFiltersService } from '@memry/app-core/saved-filters'
+import { createSavedFiltersService } from '@memry/app-core/saved-filters'
+import type { SearchReasonsService, SearchStats } from '@memry/app-core/search-tools'
 import {
   createSearchReasonsService,
   createSearchStatsService,
   createSearchTagsService
-} from './search-tools.ts'
-import type { TagsService } from './tags.ts'
-import { createTagsService } from './tags.ts'
+} from '@memry/app-core/search-tools'
+import type { TagsService } from '@memry/app-core/tags'
+import { createTagsService } from '@memry/app-core/tags'
 import type { SyncService } from './sync.ts'
 import { createSyncService } from './sync.ts'
-import type { TaskRecord, TasksService } from './tasks.ts'
-import { createTasksService } from './tasks.ts'
+import type { TaskRecord, TasksService } from '@memry/app-core/tasks'
+import { createTasksService } from '@memry/app-core/tasks'
 import type { TemplatesService } from './templates.ts'
 import { createTemplatesService } from './templates.ts'
 import type { VersionsService } from './versions.ts'
