+++ apps/cli/src/app-core/note-files.ts
@@ -5,9 +5,9 @@ import { saveCanonicalNote } from '@memry/domain-notes'
 import { getNoteMetadataByPath } from '@memry/storage-data'
 import type { FileType } from '@memry/shared/file-types'
 import { replaceWikiLinks } from '@memry/shared/wiki-target'
-import { createId } from './ids.ts'
+import { createId } from '@memry/app-core/ids'
 import type { DataDb } from './database.ts'
-import { parseMarkdownNote } from './markdown.ts'
+import { parseMarkdownNote } from '@memry/app-core/markdown'
 import { normalizePath, safeFilename, type VaultConfig } from './paths.ts'
 import type { NotesService } from './notes.ts'
