+++ apps/cli/src/app-core/paths.test.ts
