+++ apps/cli/src/app-core/paths.ts
