+++ apps/cli/src/app-core/properties.ts
