+++ apps/cli/src/app-core/sync.ts
