+++ packages/app-core/src/bookmarks.ts
@@ -1,7 +1,7 @@
 import { and, asc, eq } from 'drizzle-orm'
 import { bookmarks } from '@memry/db-schema/data-schema'
 import { bookmarkSyncId } from '@memry/contracts/bookmark-types'
-import type { DataDb } from './database.ts'
+import type { DrizzleDb as DataDb } from '@memry/db-schema/drizzle-db'
 
 export interface BookmarkRecord {
   id: string
