+++ packages/app-core/src/calendar.ts
@@ -21,7 +21,7 @@ import {
   type CalendarReminders,
   type CalendarVisibility
 } from '@memry/db-schema/data-schema'
-import type { DataDb } from './database.ts'
+import type { DrizzleDb as DataDb } from '@memry/db-schema/drizzle-db'
 import { createId } from './ids.ts'
 
 export interface CalendarEventRecord {
