+++ packages/app-core/src/reminders.ts
@@ -1,6 +1,6 @@
 import { asc, eq } from 'drizzle-orm'
 import { reminders } from '@memry/db-schema/data-schema'
-import type { DataDb } from './database.ts'
+import type { DrizzleDb as DataDb } from '@memry/db-schema/drizzle-db'
 import { createId } from './ids.ts'
 import type { ReminderTargetType, ReminderStatus } from '@memry/contracts/reminder-types'
