+++ packages/app-core/src/tasks.ts
@@ -1,6 +1,6 @@
 import { asc, eq, inArray } from 'drizzle-orm'
 import { projects, statuses, tasks, taskNotes, taskTags } from '@memry/db-schema/data-schema'
-import type { DataDb } from './database.ts'
+import type { DrizzleDb as DataDb } from '@memry/db-schema/drizzle-db'
 import { createId } from './ids.ts'
 
 export interface TaskRecord {
